add types file to thorchainSwapper
