+++ packages/swapper/src/swappers/thorchain/types.ts
@@ -1,0 +1,15 @@
+export type MidguardResponse = {
+  asset: string
+  assetDepth: string
+  assetPrice: string
+  assetPriceUSD: string
+  liquidityUnits: string
+  poolAPY: string
+  runeDepth: string
+  status: string
+  synthSupply: string
+  synthUnits: string
+  units: string
+  volume24h: string
+}
+
